Permissions applied at the row-level (for good measure)
Abstracted perm code and now also enforcing on row objects as well as on
the ResultSets. These permissions should never need to be enforced in
practice because the interfaces already only show rows the user is
allowed to access. These are just for the sake of completeness and to
protect data in the case of attempts to access direct CRUD actions. Even
then, this is mostly redundant.