Misc changes to schema
Added auto_inc column to Contact and updated the FK to use it instead of
"number", and added a unique constraint instead to enforce only one
contact number *per phone* instead of global. This allows multiple
contacts with the same number to exist in different phones, and should
have always been this way from the begining...

Also changed the 'phone_id' FKs to cascade to both contacts and
messages. This allows the user to delete a phone and automatically have
the messages and contacts in contains deleted, too.